mem: print stats

+++ src/mem/dram_ctrl.hh
@@ -1185,7 +1185,7 @@ class DRAMCtrl : public QoS::MemCtrl
 
   public:
     const AddrRange regsMap;
-    uint8_t regs[65536];
+    uint64_t regs[65536];
     DRAMCtrl(const DRAMCtrlParams* p);
 
     DrainState drain() override;
